refactor: move SuccessMediaItemsSection to shared UI components
